+++ packages/transitions/src/index.ts
@@ -9,4 +9,4 @@ export {
 	TransitionTiming,
 } from './types.js';
 // Hooks
-export {TransitionState, useTransitionProgress} from './context.js';
+export {TransitionState, useTransitionProgress} from './none.js';
